% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.pc_prediction}
\alias{print.pc_prediction}
\title{Print Method for pc_prediction}
\usage{
\method{print}{pc_prediction}(x, ...)
}
\arguments{
\item{x}{A pc_prediction object}

\item{...}{Additional arguments passed to print}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for pc_prediction
}
\keyword{internal}
