% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.summary.pc_cv}
\alias{print.summary.pc_cv}
\title{Print Method for pc_cv Summary}
\usage{
\method{print}{summary.pc_cv}(x, ...)
}
\arguments{
\item{x}{A summary.pc_cv object}

\item{...}{Additional arguments passed to print}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for pc_cv Summary
}
