% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.summary.pc_effect}
\alias{print.summary.pc_effect}
\title{Print Method for pc_effect Summary}
\usage{
\method{print}{summary.pc_effect}(x, ...)
}
\arguments{
\item{x}{A summary.pc_effect object}

\item{...}{Additional arguments passed to print}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for pc_effect Summary
}
