% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_params.R
\name{print.summary.pc_params}
\alias{print.summary.pc_params}
\title{Print method for summary(pc_params)}
\usage{
\method{print}{summary.pc_params}(x, ...)
}
\arguments{
\item{x}{An object of class "summary.pc_params"}

\item{...}{Passed to underlying print methods}
}
\description{
Print method for summary(pc_params)
}
