% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.summary.pc_state}
\alias{print.summary.pc_state}
\title{Print Method for pc_state Summary}
\usage{
\method{print}{summary.pc_state}(x, ...)
}
\arguments{
\item{x}{A summary.pc_state object}

\item{...}{Additional arguments passed to print}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for pc_state Summary
}
